backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
pydantic
python-dotenv
google-generativeai
openai
pytest
pytest-asyncio
httpx
